Convert argument of lambertw with float()
lambertw is a wrapper function. The commit coverts the first argument
x to a float with float() before passing it to the helper function. This
does the correct thing for subtypes of AbstractFloat and Integer with
no performance penalty. It also allows ForwardDiff.derivative to work if x is
of type Int, although the returned type is Float16. If x is of type
BigInt, the value returned by derivative is of type BigFloat.