Change API regarding convergence information (#33)
There is now a single function `lambertw` for computing the Lambert W
function. It takes a keyword argument, `info`. If `info` is false, the
default, then only the result of computation is returned. If it is `true`
then a triple giving the result and info on convergence is returned.

In neither case is a warning or error explicitly raised.